tickets/src/models/ticket.ts
import mongoose from "mongoose";
import { updateIfCurrentPlugin } from "mongoose-update-if-current";

//describe the properties to create a new ticket
interface TicketAttrs {
  title: string;
  price: number;
  userId: string;
}

//describe the properties of the ticket Documents must have to
//add to attrs in  the future
interface TicketDoc extends mongoose.Document {
  title: string;
  price: number;
  userId: string;
  version: number;
  orderId?: string; // optional
  // add new properties here for example createdAt
}

interface TicketModel extends mongoose.Model<TicketDoc> {
  build(attrs: TicketAttrs): TicketDoc;
}

const ticketSchema = new mongoose.Schema(
  {
    title: {
      type: String,
      //it is String not string because we here use
      //the global string constructor in JS
      required: true,
    },
    price: {
      type: Number,
      required: true,
    },
    userId: {
      type: String,
      required: true,
    },
    orderId: {
      type: String,
    },
  },
  {
    toJSON: {
      transform(doc, ret) {
        ret.id = ret._id;
        delete ret._id;
      },
    },
  }
);

ticketSchema.set("versionKey", "version");
// change the version key from the default __v to version
ticketSchema.plugin(updateIfCurrentPlugin);

ticketSchema.statics.build = (attrs: TicketAttrs) => {
  return new Ticket(attrs);
};

const Ticket = mongoose.model<TicketDoc, TicketModel>("Ticket", ticketSchema);

export { Ticket };
